Import autopager-0.1.3.0-fx.xpi with extracted autopager.jar
Compatibility: Firefox 2.0 - 2.0.0.*

Add a guid to each configed site so that when you update the site pattarn you will still keep it update. Add a quick load option. It will try to merge pages while the next page is not totally loaded(Such as some pictures not loaded yet). It's not safe for all sites. You can try it for some sites if it not work disable it.